refactor: HelpRequest 수정
-게시글 작성자 컬럼 추가
-게시글 존재여부 확인 메소드 추가
-컨트롤러 내부,외부 분리
